Adding headers options to GET
Adding the ability to send HTTP headers on GET and DELETE requests.
This change is local just to those methods and does not provide a way
to set headers across all requests at the moment.  I'm open to
discussing a global header approach but for now I wanted to keep it
simple.

closes #38, closes #25, closes #19