Fixed curl default options for headers being ignored
Methods put, post, patch, get, delete did provide an $header parameter, which (in all cases) overwrites the default options set in public $curl_opts, even if values are not present in the $header parameter. This has been changed so that the parameter data is only merged into the default parameters.